refactor: extract RegisterForm component and update RegisterPage to use it
